feat: Added the project stub provided by the Udacity team
The idea is to use the forked puckel / docker-airflow repo to first test the soulution locally, and then to deploy
the project to AWS and test it there as well. My plan is to follow the same path for the Capstone project later on.

This commit doesn`t solve any problem per se, it`s a starting point for what`s to come.

Resolves : #1